products/pyproject.toml
[tool.poetry]
name = "users"
version = "0.1.0"
description = ""
authors = ["bilalmk <bilalmk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
sqlmodel = "^0.0.19"
python-dotenv = "^1.0.1"
aiokafka = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
